Refactor config file name and setup process for destinations
